assignments/day-planner-agents/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
autogen==0.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
joblib==1.4.2
lightgbm==4.5.0
numpy==1.26.4
nvidia-nccl-cu12==2.23.4
openai==1.56.1
packaging==24.2
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xgboost==2.1.3
